+++ ubelt/__init__.py
@@ -37,7 +37,7 @@ AutogenInit:
 #     * rsync
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-__version__ = '0.8.0'
+__version__ = '0.8.1'
 
 __submodules__ = [
     'util_arg',